requirements.txt
camelot-py==0.11.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
distro==1.8.0
et-xmlfile==1.1.0
JPype1==1.4.1
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
pycparser==2.21
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pypdf==3.16.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tabula-py==2.8.2
tabulate==0.9.0
tzdata==2023.3
PyPDF2==2.0